Manually inserting type into type column.
Rails 3.2.1 required that we change Ticket to a module and include it
instead of it being an abstract_class.  This then caused the problem
where the service_ticket class was not inserting the class name into
the type column, which was raising a PG exception due to the table
having a not null constraint on this column.

We might not need this column at all, but need tests to confirm that
we don't need it.